requirements.txt
aiofiles==22.1.0
anyio==3.6.2
APScheduler==3.9.1
arrow==1.2.3
autopep8==2.0.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cookiecutter==1.7.3
fastapi==0.79.1
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
logger==1.4
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
nb-cli==0.6.7
nonebot-adapter-onebot==2.1.4
nonebot-plugin-apscheduler==0.2.0
nonebot2==2.0.0rc1
numpy==1.23.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
Pillow==9.3.0
playwright==1.27.1
poyo==0.5.0
prompt-toolkit==3.0.31
pycodestyle==2.9.1
pydantic==1.9.2
pyee==8.1.0
pyfiglet==0.8.post1
Pygments==2.13.0
pygtrie==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
retrying==1.3.3
rfc3986==1.5.0
rich==12.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.10.2
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.3
yarl==1.8.1
